long -> integer

+++ dwsys/FileInMemory.h
@@ -2,7 +2,7 @@
 #define _FileInMemory_h_
 /* FileInMemory.h
  *
- * Copyright (C) 2011-2012,2015 David Weenink, 2015 Paul Boersma
+ * Copyright (C) 2011-2017 David Weenink, 2015 Paul Boersma
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -24,7 +24,7 @@
 Thing_define (FileInMemory, Daata) {
 	char32 *d_path;
 	char32 *d_id;
-	long d_numberOfBytes;
+	integer d_numberOfBytes;
 	char *d_data;
 	bool ownData;
 
@@ -38,13 +38,13 @@ Thing_define (FileInMemory, Daata) {
 
 autoFileInMemory FileInMemory_create (MelderFile file);
 
-autoFileInMemory FileInMemory_createWithData (long numberOfBytes, const char *data, const char32 *path, const char32 *id);
+autoFileInMemory FileInMemory_createWithData (integer numberOfBytes, const char *data, const char32 *path, const char32 *id);
 
 void FileInMemory_dontOwnData (FileInMemory me);
 
 void FileInMemory_setId (FileInMemory me, const char32 *newId);
 
-void FileInMemory_showAsCode (FileInMemory me, const char32 *name, long numberOfBytesPerLine);
+void FileInMemory_showAsCode (FileInMemory me, const char32 *name, integer numberOfBytesPerLine);
 
 Collection_define (FileInMemorySet, SortedSetOf, FileInMemory) {
 	int d_sortKey;
@@ -57,16 +57,16 @@ Collection_define (FileInMemorySet, SortedSetOf, FileInMemory) {
 
 autoFileInMemorySet FileInMemorySet_createFromDirectoryContents (const char32 *dirpath, const char32 *file);
 
-void FileInMemorySet_showAsCode (FileInMemorySet me, const char32 *name, long numberOfBytesPerLine);
+void FileInMemorySet_showAsCode (FileInMemorySet me, const char32 *name, integer numberOfBytesPerLine);
 
-void FileInMemorySet_showOneFileAsCode (FileInMemorySet me, long index, const char32 *name, long numberOfBytesPerLine);
+void FileInMemorySet_showOneFileAsCode (FileInMemorySet me, integer index, const char32 *name, integer numberOfBytesPerLine);
 
-long FileInMemorySet_getIndexFromId (FileInMemorySet me, const char32 *id);
+integer FileInMemorySet_getIndexFromId (FileInMemorySet me, const char32 *id);
 
 autoStrings FileInMemorySet_to_Strings_id (FileInMemorySet me);
 
-char * FileInMemorySet_getCopyOfData (FileInMemorySet me, const char32 *id, long *numberOfBytes);
+char * FileInMemorySet_getCopyOfData (FileInMemorySet me, const char32 *id, integer *numberOfBytes);
 
-const char * FileInMemorySet_getData (FileInMemorySet me, const char32 *id, long *numberOfBytes);
+const char * FileInMemorySet_getData (FileInMemorySet me, const char32 *id, integer *numberOfBytes);
 
 #endif // _FileInMemory_h_
